pytester: isolate runpytest_subprocess from host's IO encoding
To ensure that pytester subprocess tests are not affected by host system
IO encoding (as inprocess tests are isolated, thanks to pytest's
capturing), force its stdout/stderr to a known good encoding (UTF-8).

The alternative to this approach is, instead of trying to force the
subprocess's encoding, try to adapt to it during decoding. Practically,
it would mean trying to decode with `locale.getpreferredencoding()` or
such. I opted against it in order to promote consistency across
machines, and for consistency with inprocess tests, even though it makes
the tests less realistic on the host machine.